conditional turn lookup, reuse timezone validation from extract-conditionals
