+++ include/extractor/edge_based_graph_factory.hpp
@@ -58,7 +58,7 @@ struct TurnIndexBlock
 };
 #pragma pack(pop)
 static_assert(std::is_trivial<TurnIndexBlock>::value, "TurnIndexBlock is not trivial");
-static_assert(sizeof(TurnIndexBlock) < 24, "TurnIndexBlock is not packed correctly");
+static_assert(sizeof(TurnIndexBlock) == 12, "TurnIndexBlock is not packed correctly");
 } // ns lookup
 
 struct NodeBasedGraphToEdgeBasedGraphMappingWriter; // fwd. decl
