+++ include/extractor/restriction_parser.hpp
@@ -41,14 +41,15 @@ class ScriptingEnvironment;
 class RestrictionParser
 {
   public:
-    RestrictionParser(ScriptingEnvironment &scripting_environment);
-    boost::optional<std::vector<InputRestrictionContainer>> TryParse(const osmium::Relation &relation, const bool parse_conditional) const;
+    RestrictionParser(ScriptingEnvironment &scripting_environment, bool parse_conditionals);
+    std::vector<boost::optional<InputRestrictionContainer>> TryParse(const osmium::Relation &relation) const;
 
   private:
     bool ShouldIgnoreRestriction(const std::string &except_tag_string) const;
 
     std::vector<std::string> restrictions;
     bool use_turn_restrictions;
+    bool parse_conditionals;
 };
 }
 }
