+++ src/extractor/scripting_environment_lua.cpp
@@ -498,7 +498,7 @@ void Sol2ScriptingEnvironment::ProcessElements(
         [&](const tbb::blocked_range<std::size_t> &range) {
             ExtractionNode result_node;
             ExtractionWay result_way;
-            boost::optional<std::vector<InputRestrictionContainer>> result_res;
+            std::vector<boost::optional<InputRestrictionContainer>> result_res;
             auto &local_context = this->GetSol2Context();
 
             for (auto x = range.begin(), end = range.end(); x != end; ++x)
@@ -526,13 +526,12 @@ void Sol2ScriptingEnvironment::ProcessElements(
                     resulting_ways.push_back(std::make_pair(x, std::move(result_way)));
                     break;
                 case osmium::item_type::relation:
-                    // TODO unhardcode boolean
-                    result_res.reset();
+                    result_res.clear();
                     result_res = restriction_parser.TryParse(
-                        static_cast<const osmium::Relation &>(*entity), true);
-                    if (result_res)
+                        static_cast<const osmium::Relation &>(*entity));
+                    if (result_res.size() > 0)
                     {
-                        for (const InputRestrictionContainer &r : *result_res)
+                        for (boost::optional<InputRestrictionContainer> r : result_res)
                         {
                             resulting_restrictions.push_back(r);
                         }
