+++ src/tools/extract.cpp
@@ -50,7 +50,13 @@ return_code parseArguments(int argc, char *argv[], extractor::ExtractorConfig &e
         boost::program_options::bool_switch(&extractor_config.use_metadata)
             ->implicit_value(true)
             ->default_value(false),
-        "Use metada during osm parsing (This can affect the extraction performance).");
+        "Use metadata during osm parsing (This can affect the extraction performance).")(
+        "parse-conditional-restrictions",
+        boost::program_options::value<bool>(&extractor_config.parse_conditionals)
+            ->implicit_value(true)
+            ->default_value(false),
+        "Save conditional restrictions found during extraction to disk for use "
+        "during contraction");
 
     bool dummy;
     // hidden options, will be allowed on command line, but will not be
